Small changes in a path logic for the Reinvent extension
