--- src/regex.smi
@@ -1,30 +1,0 @@
-CCCn1c(=O)c2nc(-c3ccc(/C=C/C(=O)O)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(C(=O)O)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2[nH]c(-c3ccc(OCC(=O)NCCN)cc3)nc2n(CCC)c1=O
-Nc1nc2nn(CCc3ccccc3)cc2c2nc(-c3ccco3)nn12
-CCCn1c(=O)c2nc(-c3ccc(/C=C/C(=O)NN4C(=O)C(C)=C(C)C4=O)cc3)[nH]c2n(CCC)c1=O
-Cn1cc2c(nc(NC(=O)Cc3ccccc3)n3nc(-c4ccco4)nc23)n1
-O=C(Cc1cccc2ccccc12)Nc1nc2nn(CCCc3ccccc3)cc2c2nc(-c3ccco3)nn12
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(Cl)cc4)cc3)[nH]c2n(CCC)c1=O
-CC(C)CCn1cc2c(nc(NC(=O)Cc3ccccc3)n3nc(-c4ccco4)nc23)n1
-CC(C)CCn1cc2c(nc(NC(=O)COc3ccccc3)n3nc(-c4ccco4)nc23)n1
-O=C(Cc1ccccc1)Nc1nc2nn(CCCc3ccccc3)cc2c2nc(-c3ccco3)nn12
-CC(C)CCn1cc2c(nc(NC(=O)Cc3cccc4ccccc34)n3nc(-c4ccco4)nc23)n1
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(C(=O)OC)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4cccc(C(C)=O)c4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(C(N)=O)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(I)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccccc4C(C)=O)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)NCc4ccccc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)NN4C(=O)C(C)=C(C)C4=O)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccccc4)cc3)[nH]c2n(CCC)c1=O
-Cn1cc2c(nc(NC(=O)Nc3ccccc3)n3nc(-c4ccco4)nc23)n1
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(F)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(C(C)=O)cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)O)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc([N+](=O)[O-])cc4)cc3)[nH]c2n(CCC)c1=O
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(C(F)(F)F)cc4)cc3)[nH]c2n(CCC)c1=O
-O=C(O)/C=C/c1ccc(-c2nc3c(=O)n(CC4CCCCC4)c(=O)n(CC4CCCCC4)c3[nH]2)cc1
-CCCn1c(=O)c2nc(-c3ccc(OCC(=O)Nc4ccc(Br)cc4)cc3)[nH]c2n(CCC)c1=O
-C=CCn1c(=O)c2nc(-c3ccc(OCC(=O)O)cc3)[nH]c2n(CC=C)c1=O
-O=C(COc1ccccc1)Nc1nc2nn(CCCc3ccccc3)cc2c2nc(-c3ccco3)nn12
